start of some common input validation.  overall less code.
added get_paths interface.  first path is returned w/ size/device
verified.  the second returned path (if available) is not, but it's
only returned if the device was known to be up in the past 15 seconds.
should be enough of a guarantee (for now).  we'll see.  the idea is
the client will use the 1st one in time, and at least have the second
one to fall back to if the 1st guy really does crash in the sub-second
meantime.

fixed the deleter job.  wasn't prepending the MOG_ROOT, so wasn't
deleting files.  (only from db before)

Whitaker... your turn.  :-)


git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk@6 f67b2e87-0811-0410-a7e0-dd94e48410d6