Add three new commands to mogilefsd:  CREATE_DOMAIN, CREATE_CLASS, UPDATE_CLASS
These correspond to the new MogileFS methods: create_domain, create_class, update_class

These give you the ability to use a MogileFS object to create domains and
classes on the fly.  You are also given the ability to update the minimum
replica count of an existing class in a domain using the update_class method.

$MogileFS->create_domain( $domain_name );
$MogileFS->create_class( $domain_name, $class_name, $minimum_replica_count );
$MogileFS->update_class( $domain_name, $class_name, $minimum_replica_count );


git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk/server@60 f67b2e87-0811-0410-a7e0-dd94e48410d6