MogileFSd NEXT-GENERATION.  (mogilefsd++ per Michael.)
** THIS IS NOT GUARANTEED TO BE STABLE YET, IN TESTING STILL **

This is a big overhaul of how mogilefsd works.  We've flipped it on its
head, as it were, and made what was the "frontend" job now be the main part
of the server.

Because of this, you can how give commands to mogilefsd simply by telnetting
to it as if you were a client.  (Type 'help' for help.)

You can now adjust job levels on the fly, watch errors that would have been
sent to the console, send commands to clients, see how long commands are
taking (user time, effectively), and lots of other little things like
seeing how many queries the system has performed.

Again, this is still in testing and is not stable.


git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk/server@91 f67b2e87-0811-0410-a7e0-dd94e48410d6