Add new functionality to mogilefsd: alternate zones.
Basically, you can now define two IPs for a host: the primary and the
alternate.  You can also define a netmask (in standard 0.0.0.0/0 format)
which, if matched, means that all functionality should use the alternate
zone instead of the primary.

This has uses in installations with two networks--you can have a primary IP
defined on one network and the secondary IP on the second and have it use
the secondary IP if the client comes in over that network.

You can also force responses to use the alternate zone by specifying an
argument "zone=alt" in any command.


git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk/server@132 f67b2e87-0811-0410-a7e0-dd94e48410d6