Add to mogilefsd: unreachable_fids table.
This causes the replicators to be more intelligent when they determine that
they can't get to a file for replication.

Instead of spinning around and trying the same fids over and over, they
now mark them as having failed for 60 seconds (in the case of a simple
transient failure) or 15 minutes (in the case of a more permanent failure
like a 404).

In the 15 minute case, we also insert the fid into a table named
unreachable_fids which clients will (eventually) be able to query through
the interface and get a list of files that we don't think exist anymore.

It will then be up to the client application to actually handle these, as
there is no way for MogileFS to know what should be done with these files.


git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk@146 f67b2e87-0811-0410-a7e0-dd94e48410d6