First stab at plugin support for mogilefsd, modelled after how Perlbal does it.  Still a bunch of work that needs to be done to support configuration of plugins and such.
Also include a FilePaths plugin which has initial support for supporting
creating a directory style heirarchy.  Works using the plugin system and
a couple of hooks to maintain the data structure in the database.

To come: document describing how the generic/global plugin system works, and
how to write one to run within the tracker.

This commit also fixes some minor bugs with undef being printed/used in
the cases of creating new classes, etc.




git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk/server@323 f67b2e87-0811-0410-a7e0-dd94e48410d6