Fix state_change() bug that could make file_to_delete_later grow unnecessarily (Radu Greab)
http://lists.danga.com/pipermail/mogilefs/2007-September/001273.html

	From: 	  Radu Greab
	Subject:  [PATCH] bug: some children do not receive state changes notifications
	Date:     September 13, 2007 12:49:03 AM PDT


There is a bug inside state_change() from MogileFS::ProcManager: the
cache of known states is not updated for the child that sends a state
change. Due to this some children may not be notified of future state
changes because their cached known state is wrong.

In the setup I found the above bug the load on the DB was growing
constantly due to a growing file_to_delete_later table.

The sequence of events was:
- the delete job received initially a notification that host X is
reachable and this state was cached by ProcManager.
- sometimes connections to the storage nodes timed out and the delete job
broadcasted a host X is unreachable message. The ProcManager changed the
known state cache for all children except for the delete job. The cache
for the delete job continue to says that host X is reachable.
- later, the monitor job broadcasted that host X is reachable again, but
this state change was not broadcasted to the delete job because the
cache entry already says that the host X is reachable. So the delete job
couldn't advance because its internal state said that host X was
unreachable and it didn't receive any notification from the parent that
host X is reachable again.

The fix is simple, also update the known state cache for the child that
sent the message.


git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk@1120 f67b2e87-0811-0410-a7e0-dd94e48410d6