Update zone_* settting keys to be lists of netmask
This is bad for amyways, I'm updating code in the core to follow a
change in a plugin, but I'll have to fix it later.

The new Network plugin uses a Net::Patricia (trie) to determine which
netblock a particular address belongs to. So we can support multiple
netblocks per network zone now.


git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk@1180 f67b2e87-0811-0410-a7e0-dd94e48410d6