Use gz* not deflate functions in mogtool inject
When the --gzip option is passed to mogtool inject, it uses the
deflate algorithm from Compress::Zlib, which produces different
output to the gz functions.

The gz functions are actually more useful because the original
file can be recovered with 'zcat'.

Change mogtool from use of deflate to use of gz functions.

Signed-off-by: Nick Andrew <nick@nick-andrew.net>