add new worker: job_master
could use a little cleanup, but works fine.

this new job class is meant to support the internal queue system.
it monitors the queues, and has a small bit of code for updating them.

this does break convention a bit. All of the code for a particular worker
isn't contained in the worker anymore, if they need to use a queue for work.
I'm fine with this tradeoff, since we're likely to have a lot of common queue
manager code in job_master.

git-svn-id: http://code.sixapart.com/svn/mogilefs/trunk@1247 f67b2e87-0811-0410-a7e0-dd94e48410d6